+++ frontend/pages/index.vue
@@ -3,6 +3,7 @@
     <section>
       <b-container class="main py-5 dashboard">
         <Chain />
+        <EDPBanner />
         <Search />
         <div class="row">
           <div class="col-md-6 mb-4">
@@ -67,6 +68,7 @@ import LastExtrinsics from '@/components/LastExtrinsics.vue'
 import LastEvents from '@/components/LastEvents.vue'
 import Chain from '@/components/Chain.vue'
 import Search from '@/components/Search.vue'
+import EDPBanner from '@/components/EDPBanner.vue'
 // import Top10Rich from '@/components/Top10Rich.vue'
 import { network } from '@/frontend.config.js'
 
@@ -78,6 +80,7 @@ export default {
     LastEvents,
     Chain,
     Search,
+    EDPBanner,
     // Top10Rich,
   },
   data: () => {