+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Cerestats
 
 ## vNext
-- ...
+- Added UI to show EDP banner to homepage.
 
 ## v0.1.0
 - Adapted for Cere Network.